+++ Sources/StreamChat/Controllers/ChannelListController/ChannelListController.swift
@@ -226,7 +226,9 @@ public class ChatChannelListController: DataController, DelegateCallable, DataSt
         ) { [weak self] result in
             switch result {
             case let .success(channels):
-                self?.state = .remoteDataFetched
+                self?.channelListObserver.refreshItems { [weak self] in
+                    self?.state = .remoteDataFetched
+                }
                 self?.hasLoadedAllPreviousChannels = channels.count < limit
                 self?.callback { completion?(nil) }
             case let .failure(error):